cte_engine/requirements.txt
fastapi
uvicorn
motor
redis
qdrant-client
google-generativeai
langgraph
langchain-core
python-dotenv
httpx
pydantic
pydantic-settings
streamlit
watchdog